# packages/backend/src/api/swagger.yaml
# @format

swagger: '2.0'
info:
    description: 'This is a bitcoin HD wallet server.'
    version: '1.0.0'
    title: 'Bitcoin HD Wallet'
    termsOfService: 'http://swagger.io/terms/'
    contact:
        email: 'joez@qq.com'
    license:
        name: 'Apache 2.0'
        url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: 'localhost:3000'
basePath: '/v1'
tags:
    - name: 'wallet'
      description: 'Bitcoin HD Wallet'
schemes:
    - 'http'
paths:
    /wallet/seed:
        get:
            tags:
                - 'wallet'
            summary: 'generate a random BIP39 Seed'
            operationId: 'generateSeed'
            produces:
                - 'application/json'
            parameters: []
            responses:
                200:
                    description: 'successful operation'
                    schema:
                        $ref: '#/definitions/SeedCreationResponse'
                500:
                    description: 'Internal Server Error'
            x-swagger-router-controller: 'WalletController'
    /wallet/P2WPKHAddress:
        post:
            tags:
                - 'wallet'
            summary: 'generate HD segwit Pay-to-Public-Key-Hash bitcoin address'
            operationId: 'generateP2WPKHAddress'
            produces:
                - 'application/json'
            parameters:
                - in: 'body'
                  name: 'body'
                  description: 'seed and path for HD wallet to generate segWit bitcoin address'
                  required: true
                  schema:
                      $ref: '#/definitions/AddressCreationRequest'
            responses:
                200:
                    description: 'successful operation'
                    schema:
                        $ref: '#/definitions/AddressCreationResponse'
                400:
                    description: 'Invalid seed or path supplied'
                500:
                    description: 'Internal Server Error'
            x-swagger-router-controller: 'WalletController'
    /wallet/P2PKHAddress:
        post:
            tags:
                - 'wallet'
            summary: 'generate HD legacy Pay-to-Public-Key-Hash bitcoin address'
            operationId: 'generateP2PKHAddress'
            produces:
                - 'application/json'
            parameters:
                - in: 'body'
                  name: 'body'
                  description: 'seed and path for HD wallet to generate Legacy bitcoin address'
                  required: true
                  schema:
                      $ref: '#/definitions/AddressCreationRequest'
            responses:
                200:
                    description: 'successful operation'
                    schema:
                        $ref: '#/definitions/AddressCreationResponse'
                400:
                    description: 'Invalid seed or path supplied'
                500:
                    description: 'Internal Server Error'
            x-swagger-router-controller: 'WalletController'
    /wallet/P2SHAddress:
        post:
            tags:
                - 'wallet'
            summary: 'generate Pay-to-ScriptHash bitcoin address'
            operationId: 'generateP2SHAddress'
            produces:
                - 'application/json'
            parameters:
                - in: 'body'
                  name: 'body'
                  description: 'the public keys of all the participants and the amount of signatures required to release the coins'
                  required: true
                  schema:
                      $ref: '#/definitions/P2SHAddressCreationRequest'
            responses:
                200:
                    description: 'successful operation'
                    schema:
                        $ref: '#/definitions/AddressCreationResponse'
                400:
                    description: 'Invalid parameters supplied'
                500:
                    description: 'Internal Server Error'
            x-swagger-router-controller: 'WalletController'
definitions:
    AddressCreationRequest:
        type: 'object'
        properties:
            seed:
                type: 'string'
                description: 'hex string to generate master key of HD wallet'
                example: 'f4f0cda65a9068e308fad4c96e8fe22213dd535fe7a7e91ca70c162a38a49aaacfe0dde5fafbbdf63cf783c2619db7174bc25cbfff574fb7037b1b9cec3d09b6'
            path:
                type: 'string'
                description: 'the path of child node in HD wallet, the formate follow the stand of bip32'
                example: 'm/84/0/0/0/6'
    AddressCreationResponse:
        type: 'object'
        properties:
            address:
                type: 'string'
                example: 'bc1q54xjmr2fmwc5hewjrj7dddaxtj6purftrnp7jp'
    P2SHAddressCreationRequest:
        type: 'object'
        properties:
            pubkeys:
                type: 'array'
                description: 'the list of public keys'
                items:
                    type: 'string'
                example:
                    - '022bfa20bf739d5072320c128ca6734c2e177dff913b9a4d39812acddbeb5802c5'
                    - '0391eec60a617e2beccc684b617d36a26f08e2b2b64f1e72e1d8a8ec632a8112b6'
                    - '035403ded32575dd0170b1203ef8f91be3b735fdeeaaf1302840e2f63dad7191b9'
                    - '043acc2157d2ed9659dcddce97280fa70bae74e1c757f4457b7296aba4f81141a37cb5628d9cff2f1be06eb528de4eaa3ee51d93a979fe854b715dc40bfd61b236'
            m:
                type: 'number'
                description: 'the amount of signatures required to release the coin'
                example: 2
    SeedCreationResponse:
        type: 'object'
        properties:
            seed:
                type: 'string'
                description: 'seed in hex string'
                example: '0ab14e16fab215521bde35d582015d92cf53d553f92754c8377b3adcd7370dfe0a51ea8a448297447916daf55e0d7ee52155c2d2b2ebc2acb36372021917d30b'
externalDocs:
    description: 'Find out more about Swagger'
    url: 'http://swagger.io'